Refactored some of the indexing functions into the utils directory
