+++ halotools/empirical_models/phase_space_models/subhalo_based_models/subhalo_selection_kernel.py
@@ -5,6 +5,8 @@ from __future__ import division, print_function, absolute_import, unicode_litera
 import numpy as np
 from astropy.utils.misc import NumpyRNGContext
 
+from ....utils import calculate_first_idx_unique_array_vals, sum_in_bins
+
 __all__ = ('subhalo_indexing_array', )
 
 
@@ -360,114 +362,6 @@ def array_weave(val1, val2, mult1, mult2, testing_mode=False):
     result[val1_mask] = val1
     result[~val1_mask] = val2
     return result, ~val1_mask
-
-
-def calculate_first_idx_unique_array_vals(sorted_array):
-    """ Given an integer array with possibly repeated entries in ascending order,
-    return the indices of the first appearance of each unique value.
-
-    Parameters
-    ----------
-    sorted_array : array
-        Integer array of host halo IDs, sorted in ascending order
-
-    Returns
-    --------
-    idx_unique_array_vals : array
-        Integer array storing the indices of the first appearance of
-        each unique entry in sorted_array
-
-    Examples
-    --------
-    >>> sorted_array = np.array((0, 0, 1, 1, 4, 8, 8, 10))
-    >>> result = calculate_first_idx_unique_array_vals(sorted_array)
-    >>> assert np.all(result == (0, 2, 4, 5, 7))
-    """
-    return np.concatenate(([0], np.flatnonzero(np.diff(sorted_array)) + 1))
-
-
-def calculate_last_idx_unique_array_vals(sorted_array):
-    """ Given an integer array with possibly repeated entries in ascending order,
-    return the indices of the last appearance of each unique value.
-
-    Parameters
-    ----------
-    sorted_array : array
-        Integer array of host halo IDs, sorted in ascending order
-
-    Returns
-    --------
-    idx_unique_array_vals : array
-        Integer array storing the indices of the last appearance of
-        each unique entry in sorted_array
-
-    Examples
-    --------
-    >>> sorted_array = np.array((0, 0, 1, 1, 4, 8, 8, 10))
-    >>> result = calculate_last_idx_unique_array_vals(sorted_array)
-    >>> assert np.all(result == (1, 3, 4, 6, 7))
-    """
-    return np.append(np.flatnonzero(np.diff(sorted_array)), len(sorted_array)-1)
-
-
-def sum_in_bins(arr, sorted_bin_numbers, testing_mode=False):
-    """ Given an array of values ``arr`` and another equal-length array
-    ``sorted_bin_numbers`` storing how these values have been binned into *Nbins*,
-    calculate the sum of the values in each bin.
-
-    Parameters
-    -----------
-    arr : array
-        Array of length *Nvals* storing the quantity to be summed
-        in the bins defined by ``sorted_bin_numbers``.
-
-    sorted_bin_numbers : array
-        Integer array of length *Nvals* storing the bin numbers
-        of each entry of the input ``arr``,
-        e.g., the result of np.digitize(arr, bins).
-        The ``sorted_bin_numbers`` array may have repeated entries but must
-        be in ascending order. That is, the subhalos whose property is
-        stored in array ``arr`` will be presumed to be pre-grouped according to,
-        for example, host halo mass, with lowest halo masses first,
-        and higher halo masses at higher indices, in monotonic fashion.
-
-    testing_mode : bool, optional
-        Boolean specifying whether input arrays will be tested to see if they
-        satisfy the assumptions required by the algorithm.
-        Setting ``testing_mode`` to True is useful for unit-testing purposes,
-        while setting it to False improves performance.
-        Default is False.
-
-    Returns
-    --------
-    binned_sum : array
-        Array of length-*Nbins* storing the sum of ``arr`` values
-        within the bins defined by ``sorted_bin_numbers``.
-
-    Examples
-    --------
-    >>> Nvals = 5
-    >>> arr = np.arange(Nvals)
-    >>> sorted_bin_numbers = np.array((1, 2, 2, 6, 7))
-    >>> result = sum_in_bins(arr, sorted_bin_numbers)
-    """
-    try:
-        assert len(arr) == len(sorted_bin_numbers)
-    except AssertionError:
-        raise ValueError("Input ``arr`` and ``sorted_bin_numbers`` must have same length")
-
-    if testing_mode is True:
-        try:
-            assert np.all(np.diff(sorted_bin_numbers) >= 0)
-        except AssertionError:
-            msg = ("Input ``sorted_bin_numbers`` array must be sorted in ascending order")
-            raise ValueError(msg)
-
-    last_idx = calculate_last_idx_unique_array_vals(sorted_bin_numbers)
-    first_entry = arr[:last_idx[0]+1].sum()
-    binned_sum = np.diff(np.cumsum(arr)[last_idx])
-    binned_sum = np.concatenate(([first_entry], binned_sum))
-    return binned_sum
 
 
 def indices_of_selected_subhalos(subhalo_hostids, subhalo_occupations, subhalo_multiplicity,
