+++ halotools/utils/__init__.py
@@ -9,3 +9,4 @@ from .table_utils import *
 from .value_added_halo_table_functions import *
 from .group_member_generator import group_member_generator
 from .crossmatch import crossmatch
+from .array_indexing_manipulations import *
